// src/infrastructure/services/star-wars-api-translation/interfaces/planets-translation.interface.ts
export interface PlanetsTranslation {
    nombre: string;
    periodo_rotacion: string;
    periodo_orbital: string;
    diametro: string;
    clima: string;
    gravedad: string;
    terreno: string;
    agua_superficial: string;
    poblacion: string;
    residentes: string[];
    peliculas: string[];
    url: string;
    creado: string;
    editado: string;
}