// src/infrastructure/services/star-wars-api-translation/interfaces/films-translation.interface.ts
export interface FilmsTranslation {
    titulo: string;
    episodio_id: number;
    rastreo_inicial: string;
    director: string;
    productor: string;
    fecha_estreno: string;
    personajes: string[];
    planetas: string[];
    naves: string[];
    vehiculos: string[];
    especies: string[];
    url: string;
    creado: string;
    editado: string;
}