// src/infrastructure/services/star-wars-api-translation/interfaces/starship-translation.interface.ts
export interface StarshipsTranslation {
    nombre: string;
    modelo: string;
    fabricante: string;
    costo_en_creditos: string;
    longitud: string;
    velocidad_maxima_atmosferica: string;
    tripulacion: string;
    pasajeros: string;
    capacidad_carga: string;
    consumibles: string;
    calificacion_hiperimpulsor: string;
    MGLT: string;
    clase_nave: string;
    pilotos: string[];
    peliculas: string[];
    creado: string;
    editado: string;
    url: string;
}