// src/infrastructure/services/star-wars-api-translation/interfaces/vehicles-translation.interface.ts
export interface VehiclesTranslation {
    nombre: string;
    modelo: string;
    fabricante: string;
    costo_en_creditos: string;
    longitud: string;
    velocidad_maxima_atmosferica: string;
    tripulacion: string;
    pasajeros: string;
    capacidad_carga: string;
    consumibles: string;
    clase_vehiculo: string;
    pilotos: string[];
    peliculas: string[];
    creado: string;
    editado: string;
    url: string;
}