// src/infrastructure/services/star-wars-api-translation/interfaces/species-translation.interface.ts
export interface SpeciesTranslation {
    nombre: string;
    clasificacion: string;
    designacion: string;
    altura_promedio: string;
    esperanza_vida_promedio: string;
    colores_cabello: string;
    colores_piel: string;
    colores_ojos: string;
    mundo_natal: string;
    lenguaje: string;
    personas: string[];
    peliculas: string[];
    url: string;
    creado: string;
    editado: string;
}